+++ src/main/java/frc/robot/Robot.java
@@ -1,24 +1,21 @@
 
 package frc.robot;
 
-import com.revrobotics.ColorSensorV3;
-import com.revrobotics.ColorMatch;
-import com.revrobotics.ColorMatchResult;
 
+
+import edu.wpi.first.wpilibj.smartdashboard.SendableChooser;
 import edu.wpi.first.wpilibj.Spark;
 import edu.wpi.first.wpilibj.TimedRobot;
 import edu.wpi.first.wpilibj.Timer;
 import edu.wpi.first.wpilibj.XboxController;
+import edu.wpi.first.wpilibj.GenericHID.Hand;
 import edu.wpi.first.wpilibj.drive.DifferentialDrive;
-import edu.wpi.first.wpilibj.smartdashboard.SendableChooser;
-import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 // import edu.wpi.first.networktables.NetworkTableEntry;
 // import edu.wpi.first.wpilibj.DriverStation;
-import edu.wpi.first.wpilibj.I2C;
 import edu.wpi.first.wpilibj.TimedRobot;
 // import edu.wpi.first.wpilibj.shuffleboard.Shuffleboard;
 // import edu.wpi.first.wpilibj.shuffleboard.ShuffleboardTab;
-import edu.wpi.first.wpilibj.util.Color;
+import edu.wpi.first.wpilibj.GenericHID;
 
 public class Robot extends TimedRobot {
   private static final String kDefaultAuto = "Default";
@@ -28,20 +25,12 @@ public class Robot extends TimedRobot {
   private final DifferentialDrive robotDrive = new DifferentialDrive(new Spark(0), new Spark(1)); //Sparks have placeholder values
   private final XboxController c_xbox = new XboxController(0); //Xbox controller port is a placeholder
   private final Timer timer = new Timer();
-  private final I2C.Port i2Port = I2C.Port.kOnboard;
-  private final ColorSensorV3 s_color = new ColorSensorV3(i2Port);
-  private ColorMatch colormatcher = new ColorMatch();
-  private final Color kRedTarget = ColorMatch.makeColor(0.54, 0.33, 0.13);
-  private final Color kGreenTarget = ColorMatch.makeColor(0.16, 0.57, 0.26); 
-  private final Color kBlueTarget = ColorMatch.makeColor(0.12, 0.40, 0.47);
-  private final Color kYellowTarget = ColorMatch.makeColor(0.31, 0.56, 0.13);
-  
+  private final Spark spin = new Spark(0); //placeholder value
   
   
   // private ShuffleboardTab dataTab = Shuffleboard.getTab("Sensor");
   // private NetworkTableEntry redEntry, greenEntry, blueEntry, irEntry;
-  Color detectedColor;
-  double IR;
+ 
   
 
 
@@ -51,13 +40,7 @@ public class Robot extends TimedRobot {
    */
   @Override
   public void robotInit() {
-    m_chooser.setDefaultOption("Default Auto", kDefaultAuto);
-    m_chooser.addOption("My Auto", kCustomAuto);
-    SmartDashboard.putData("Auto choices", m_chooser);
-    colormatcher.addColorMatch(kBlueTarget);
-    colormatcher.addColorMatch(kRedTarget);
-    colormatcher.addColorMatch(kGreenTarget);
-    colormatcher.addColorMatch(kYellowTarget);
+    
   }
 
   /**
@@ -74,43 +57,11 @@ public class Robot extends TimedRobot {
 
   //  }
   @Override
-  public void robotPeriodic() {
-    Color detectedColor = s_color.getColor();
-    IR = s_color.getIR();
-    // redEntry.setDouble(detectedColor.red);
-    // greenEntry.setDouble(detectedColor.green);
-    // blueEntry.setDouble(detectedColor.blue);
-    // irEntry.setDouble(IR);
+  public void robotPeriodic(){
+
+  }
     
-    // RGB Values
-    //blue = 0.120605 0.407227 0.472412
-    //green = 0.163086 0.573975 0.262939
-    //red = 0.542236 0.327393 0.130127
-    //yellow = 0.317871 0.556641 0.125488 
-   
-    String colorString;
-    ColorMatchResult match = colormatcher.matchClosestColor(detectedColor);
-
-
-    if (match.color == kBlueTarget){
-      colorString = "Blue";
-    } else if (match.color == kRedTarget) {
-      colorString = "Red";
-    } else if (match.color == kGreenTarget) {
-      colorString = "Green";
-    } else if (match.color == kYellowTarget) {
-      colorString = "Yellow";
-    } else {
-      colorString = "Unknown";
-    }
-    SmartDashboard.putNumber("Red", detectedColor.red);
-    SmartDashboard.putNumber("Green", detectedColor.green);
-    SmartDashboard.putNumber("Blue", detectedColor.blue);
-    SmartDashboard.putNumber("Confidence", match.confidence);
-    SmartDashboard.putString("Detected Color", colorString);
-    SmartDashboard.putNumber("IR", IR);
-}
-
+//this is a random message
   /**
    * This autonomous (along with the chooser code above) shows how to select
    * between different autonomous modes using the dashboard. The sendable
@@ -165,7 +116,9 @@ public class Robot extends TimedRobot {
 
   @Override
   public void teleopPeriodic() {
-  
+  if(c_xbox.getBumper(Hand.kLeft)){ //when left bumper is pressed, the wheel should function to spin accordingly
+    spin.set(1.0); //placeholder numbers
+  }
 
   /**
    * In addition to RGB IR values, the color sensor can also return an 
@@ -178,11 +131,11 @@ public class Robot extends TimedRobot {
    * or provide a threshold for when an object is close enough to provide
    * accurate color values.
    */
-  int proximity = s_color.getProximity();
+  
 
-  SmartDashboard.putNumber("Proximity", proximity);
 
    }
+   
 
   /**
    * This function is called periodically during test mode.
@@ -190,4 +143,5 @@ public class Robot extends TimedRobot {
   @Override
   public void testPeriodic() {
   }
+
 }
